+++ docs/changelog.rst
@@ -7,6 +7,9 @@ v0.6 (not yet released)
 * Print function and functions in christmas tree pattern
   `#62 <https://github.com/msoeken/kitty/pull/62>`_
 
+* Get runlength encoding of truth table bits
+  `#63 <https://github.com/msoeken/kitty/pull/63>`_
+
 v0.5 (October 8, 2018)
 ----------------------