+++ docs/changelog.rst
@@ -10,6 +10,9 @@ v0.6 (not yet released)
 * Get runlength encoding of truth table bits
   `#63 <https://github.com/msoeken/kitty/pull/63>`_
 
+* Constructors: ``create_from_parity``
+  `#64 <https://github.com/msoeken/kitty/pull/64>`_
+
 v0.5 (October 8, 2018)
 ----------------------